Matter: simplify ConnectivityManagerImpl
