+++ lib/matter_ext.scons
@@ -464,7 +464,7 @@ sources = [
     f"{MATTER_GLUE_DIR}/platform/bsb/CHIPCryptoPALTinyCrypt.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/CommissionableDataProviderImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/ConfigurationManagerImpl.cpp",
-    f"{MATTER_GLUE_DIR}/platform/bsb/ConnectivityManagerImpl_WIFI.cpp",
+    f"{MATTER_GLUE_DIR}/platform/bsb/ConnectivityManagerImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/DeviceInstanceInfoProviderImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/DiagnosticDataProviderImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/GenericPlatformManagerImplFuri.cpp",
