+++ lib/matter_glue/platform/bsb/CHIPDevicePlatformConfig.h
@@ -154,3 +154,5 @@
 
 // FIXME just for testing
 #define CHIP_DEVICE_CONFIG_ENABLE_PAIRING_AUTOSTART 0
+
+#define CHIP_DEVICE_CONFIG_ENABLE_WIFI_STATION 0
